==> 00_ReadMe.ipynb <==
# READ ME: detectie stoplocaties uit AIS-data

##### Onderzoeksvraag
Welke techniek kunnen we inzetten om een generiek model te ontwikkelen om de computer zelf stoplocaties uit XYT gegevens te laten detecteren? 

##### Model detectie stoplocaties AIS-data 
Het doel van dit model is het detecteren van stoplocaties op basis van AIS gegevens van schepen. Het model is het startpunt van een groter generiek model die gedrag binnen GPS gegevens kan herkennen op basis van XY (latitude, longitude) en T (tijd) gegevens. Wij wensen dat de computer zelf automatisch gedrag kan herkennen, zonder menselijke aannames. Daarnaast hopen wij dit model ook in te kunnen zetten voor andere type vraagstukken dan AIS-data (generiek).    

Gezien er sprake is van ongelabelde data, kiezen wij voor een de density-based clustertechniek DBScan. Deze techniek clustert gebieden met een hoge densititeit (punten lijken op elkaar) en scheidt van gebieden met een lagere densititeit. De gebieden met een lagere densititeit (en niet in een cluster vallen) wordt ruis genoemd. Binnen ons project zijn de clusters stoplocaties en bevat de ruis de varende punten. Hoe de DBScan clustert is afhankelijk van de features en de parameters die je meegeeft. De features die wij willen meegeven zijn gebasseerd op de XYT gegevens (bijv. afstand, snelheid, hoek en tijd tussen twee punten). De parameters die we meegeven is epsilon (radius om punt) en het minimale aantal punten binnen de epsilon.   

Gezien wij wensen een generiek en automatisch model te creëren, willen wij de computer zelf parameters laten selecteren. Om zicht te krijgen wat de solution space is voor deze parameters, hebben wij de distributies van clusters en niet clusters vergeleken met de Mann-Withney U. Waarbij een lage U waarde staat voor een groot verschil in distributies tussen de groepen en dus inzicht biedt voor het selecteren van de juiste parameters.        

##### AIS data
Samenvattend bevat AIS data gegevens over de identiteit, positie, snelheid en koers van schepen. De volgende gegevens worden verzonden: AIS transmitters verstuurd elke 2 to 10 seconden een signaal wanneer het schip vaart en elke 3 minuten wanneer het schip aan anker ligt. De main fields die worden verstuurd zijn:

    - MMSI (Maritime Mobile Service Identity) – a series of nine digits
      uniquely identifying ship stations;
    - Navigation status – “at anchor”, “under way using engine(s)”, or “not
      under command”;
    - Rate of turn – right or left, 0 to 720 degrees per minute;
    - Speed over ground – 0 to 102 knots with 0.1 knot resolution;
    - Position accuracy;
    - Longitude and Latitude – to 1/10,000 minute;
    - Course over ground – relative to true north to 0.1 degree;
    - True Heading – 0 to 359 degrees from gyro compass;
    - Timestamp – Coordinated Universal Time (UTC) time accurate to nearest
      second when this data was generated.

 Daarnaast zend de AIS transmitter additional fields:
     - IMO (International Maritime Organization) ship identification number
     - a seven digit number that remains unchanged upon transfer of the
       ship's registration to another Country;
     - International radio call sign – up to seven characters, assigned to
       the vessel by its Country of registry;
     - Vessel Name – 20 characters to represent the name of the vessel;
     - Type of ship/cargo;
     - Dimensions of ship – to nearest meter;
     - Type of positioning system – such as GPS, Differential Global
       Positioning Systems (DGPS) or Long Range Navigation (LORAN)-C;
     - Location of positioning system's antenna on-board the vessel;
     - Draught of ship – 0.1 meter to 25.5 meters;
     - Destination – max 20 characters;
     - Estimated time of arrival (ETA) at destination – UTC date hour:
       minute.
Bron: https://www.sciencedirect.com/science/article/pii/S2405535216300201

De reporting interval verhoogd als de snelheid van een schip verhoogd:

- Position report:
    - 3 min: at anchor
    - 10 sec: < 14 knopen
    - 6 sec: < 23 knopen
- Standard Class B equipment position report:
    - 3 min: < 2 knots
    - 30 sec: < 14 knots
    - 15 sec: < 23 knots
Bron: https://www.sciencedirect.com/science/article/pii/S0308597X17305535

Overige artikelen: https://www.sciencedirect.com/science/article/pii/S0888613X13000728

##### Opbouw model
We hebben het model opgedeeld in 4 verschillende scripts. 
    - Preprocessing van data --> output: CSV file
    - Runnen van DBScan met verschillende parameters --> output: CSV file
    - Visualiseren van output DBScan om de juiste parameters te selecteren --> output: data visualisatie
    - DBScan runnen met juiste parameters + output visualiseren --> output: CSV file, data visualisatie, interactieve kaart
